Make `x86-lookup-create-index' coding-system robust.
The mnemonic regexp in `x86-lookup-create-index' does not work if, for
example, the coding system is `latin-0'; this is because the temporary
buffer filled by `(call-process x86-lookup-pdftotext-program ...)' may
be not encodable in non-UTF-8 coding system, and hence will generate
spurious characters that disrupt the regexp matching.

This patch binds the `coding-system-for-read' and
`coding-system-for-write' to `utf-8' prior to `call-process' in
`x86-lookup-create-index' to make the index creation coding-system robust.